App Workflow - Home Screen
This is a refined workflow of the original. Font size reduced to 12 pt. Example of what the Home Screen could display, such as reminders to take daily dose. Tracking whether the morning dose has been logged or not, and setting a reminder for last dose before bed. Option for Push Notifications outside of the app pending.